Allow using non-refcounted records as signal parameters
Their GValue traits use const pointers for the getters to enforce
that the values are not owned by the caller and not supposed to be
modified.

For signals this then also requires to pass such records as const
pointers to signal handlers.

Fixes https://gitlab.gnome.org/bugaevc/peel/-/issues/40