Fix casting fixed-size arrays
We were using plain_cpp_type, i.e. passing strip_refs to
generate_cpp_type(), but using the resulting type as if it nothing was
stripped. Coupled with generate_cpp_type() stripping refs in some cases,
and forgetting to do it in others, this was somewhat working.

Fix generate_cpp_type() to strip refs properly, and open-code more of
the logic in generate_cast_from_c().